requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anvil-parser==0.9.0
async-timeout==4.0.2
attrs==22.1.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==37.0.1
frozendict==2.3.4
frozenlist==1.3.1
fsspec==2022.7.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
NBT==1.5.1
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
pip==22.1.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytorch-lightning==1.7.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
setuptools==61.2.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
wheel==0.37.1
yarl==1.8.1
zipp==3.8.1
